+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/Data/PackageDependency.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Studio23.SS2.UPMAssistant.Editor.Data
+{
+    public class PackageDependency
+    {
+        [JsonProperty("name")] public string Name;
+        [JsonProperty("version")] public string Version;
+    }
+}