+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/Models/PackageDependency.cs
@@ -1,8 +1,0 @@
-﻿namespace Studio23.SS2.UPMAssistant.Editor
-{
-    public class PackageDependency
-    {
-        public string name;
-        public string version;
-    }
-}