+++ Assets/Packages/com.studio23.ss2.upmassistant/CHANGELOG.md
@@ -10,3 +10,15 @@
 - dependency add and remove bug fixed.
 - refresh and delete button add.
 - tests asmdef included.
+
+
+## [1.0.3](2023-10-19)
+
+- Delete system, delete folder structure with save data
+- Create folder structure if not exist
+- Markup edit system
+- Added assembly under Runtime folder, added two folders under Runtime folder: Core, Data
+- Introduced online license fetching system
+- Licenses can be fetched, displayed in a dropdown, downloaded, saved, updated, and deleted
+- package.json is automatically updated when updating the license
+- Automatically update package.json name, company name, package version while updating in Player Settings